requirements.txt
django<1.7
django-pagedown
django-imagekit
MySQL-python
PIL
